+++ src/controllers/delete.js
@@ -82,12 +82,7 @@ class Delete {
                 });
             }],
             delete_folder: ['clear_object', (r, callback) => {
-                let folderName = _.get(this.json, 'user');
-                if (Config.get('sftp.useNewFormat', false)) {
-                    folderName = this.json.uuid;
-                }
-
-                Fs.remove(Path.join(Config.get('sftp.path', '/srv/daemon-data'), folderName), callback);
+                Fs.remove(Path.join(Config.get('sftp.path', '/srv/daemon-data'), this.json.uuid), callback);
             }],
         }, err => {
             if (err) Log.fatal(err);